change type data on produk

+++ database/migrations/2023_01_28_040821_create_produk_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('produk', function (Blueprint $table) {
             $table->id();
             $table->string('nama');
-            $table->decimal('harga');
+            $table->bigInteger('harga');
             $table->string('foto');
             $table->text('spesifikasi');
             $table->bigInteger('stok');
